refactor: rename WalletInlineBalance
