+++ src/frontend/src/lib/components/wallet/WalletInlineBalance.svelte
